rename posts -> post
