+++ src/components/model/post/PostHeader/PostHeader.module.css
