+++ src/components/model/post/Thumbnail/Thumbnail.tsx
