+++ src/components/model/post/Thumbnail/index.ts
