+++ src/components/model/post/ThumbnailDummy/ThumbnailDummy.module.css
