+++ src/components/model/post/ThumbnailDummy/ThumbnailDummy.tsx
