+++ src/components/model/post/ThumbnailDummy/index.ts
