+++ src/pages/index.tsx
@@ -3,8 +3,8 @@ import { InferGetStaticPropsType, GetStaticProps } from "next"
 import { SeoHeaders } from "components/ui/SeoHeaders"
 import { Layout } from "components/Layout/Layout"
 import { Wrapper } from "components/Wrapper/Wrapper"
-import { ArticleCard } from "components/model/posts/ArticleCard/ArticleCard"
-import { ArticleCardWrapper } from "components/model/posts/ArticleCardWrapper/ArticleCardWrapper"
+import { ArticleCard } from "components/model/post/ArticleCard/ArticleCard"
+import { ArticleCardWrapper } from "components/model/post/ArticleCardWrapper/ArticleCardWrapper"
 import { PostData } from "../components/interface/post-data"
 import { BlogPostRepository } from "../infra/blog-post-repository"
